+++ layout/src/main/java/com/vlkan/log4j2/logstash/layout/resolver/ExceptionResolver.java
@@ -42,7 +42,7 @@ class ExceptionResolver implements EventResolver {
 
                 @Override
                 EventResolver createStackTraceTextResolver(EventResolverContext context) {
-                    StackTraceTextResolver stackTraceTextResolver = new StackTraceTextResolver(context.getWriterPool());
+                    StackTraceTextResolver stackTraceTextResolver = new StackTraceTextResolver(context.getWriterCapacity());
                     return (logEvent, jsonGenerator) -> {
                         Throwable exception = logEvent.getThrown();
                         if (exception == null) {