+++ layout/src/main/java/com/vlkan/log4j2/logstash/layout/resolver/StackTraceTextResolver.java
@@ -2,27 +2,40 @@ package com.vlkan.log4j2.logstash.layout.resolver;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.vlkan.log4j2.logstash.layout.util.BufferedPrintWriter;
-import com.vlkan.log4j2.logstash.layout.util.BufferedPrintWriterPool;
+import org.apache.logging.log4j.util.Constants;
 
 import java.io.IOException;
+import java.util.function.Supplier;
 
 class StackTraceTextResolver implements StackTraceResolver {
 
-    private final BufferedPrintWriterPool writerPool;
+    private final Supplier<BufferedPrintWriter> writerSupplier;
 
-    StackTraceTextResolver(BufferedPrintWriterPool writerPool) {
-        this.writerPool = writerPool;
+    private final ThreadLocal<BufferedPrintWriter> writerRef;
+
+    StackTraceTextResolver(int writerCapacity) {
+        this.writerSupplier = () -> BufferedPrintWriter.ofCapacity(writerCapacity);
+        this.writerRef = Constants.ENABLE_THREADLOCALS
+                ? ThreadLocal.withInitial(writerSupplier)
+                : null;
     }
 
     @Override
     public void resolve(Throwable throwable, JsonGenerator jsonGenerator) throws IOException {
-        BufferedPrintWriter writer = writerPool.acquire();
-        try {
-            throwable.printStackTrace(writer);
-            jsonGenerator.writeString(writer.getBuffer(), 0, writer.getPosition());
-        } finally {
-            writerPool.release(writer);
+        BufferedPrintWriter writer = getResetWriter();
+        throwable.printStackTrace(writer);
+        jsonGenerator.writeString(writer.getBuffer(), 0, writer.getPosition());
+    }
+
+    private BufferedPrintWriter getResetWriter() {
+        BufferedPrintWriter writer;
+        if (Constants.ENABLE_THREADLOCALS) {
+            writer = writerRef.get();
+            writer.close();
+        } else {
+            writer = writerSupplier.get();
         }
+        return writer;
     }
 
 }